third_party/blink/web_tests/external/wpt/css/selectors/invalidation/user-action-pseudo-classes-in-has-expected.txt

This is a testharness.js-based test.
[FAIL] CSS Selectors Invalidation: user-action pseudo classes in :has() argument
  assert_equals: subject1 should be blue again expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
